refactor(install): add prompt_option to simplify menu options
This is similar to the bash "select" command where the text value of the selected option is populated into the a variable with the name given as the first parameter. In addition a REPLY variable is set with the numerical value.

In cases where the options are static, doing a case on the REPLY variable is probably easiest. In cases where the options may be dynamic it might make sense to do a case on the text value of the option because numerical order may vary.